requirements.txt
Django>=4.2,<5.0
djangorestframework
django-jazzmin
Pillow
psycopg2-binary>=2.9.9
drf-yasg
drf_spectacular
django-filter>=23.5
djangorestframework-simplejwt
gunicorn
django-cors-headers
whitenoise
